backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0

# Redis
redis[hiredis]>=5.2.0

# Auth
PyJWT>=2.10.0
bcrypt>=4.2.0

# HTTP Client (LLM calls)
httpx>=0.28.0

# File handling
python-multipart>=0.0.18

# Utilities
python-dotenv>=1.0.1
